#103 removed bugs from code

+++ app/src/main/java/com/conupods/OutdoorMaps/Services/LocationService.java
@@ -37,7 +37,7 @@ public class LocationService extends Service {
     private static final int RESOLVABLE_API_ERROR_REQUEST_CODE = 51;
     private static final String TAG = "LOCATION_SERVICE";
 
-    private final static long UPDATE_INTERVAL = 4 * 1000;  /* 4 secs */
+    private final static long UPDATE_INTERVAL = (long)4 * (long)1000;  /* 4 secs */
     private final static long FASTEST_INTERVAL = 2000; /* 2 sec */
 
     @Nullable
